Feedback processing phase 1/2: - [Coin list] Auto-retry on request failure - [Network] Double -> Decimal - [Network] Eliminated some excessive initializers - [Infra] Tweaked .gitignore - [Error popup] does not trigger the whole hierarchy redrawing anymore

+++ CoinGeckoDemo/CoinGeckoAPI/Sources/CoinGeckoAPI/Domain/CoinList/Coin.swift
@@ -5,6 +5,6 @@ public struct Coin: Decodable {
     public let symbol: String
     public let name: String
     public let image: URL
-    public let currentPrice: Double
-    public let priceChangePercentage24H: Double
+    public let currentPrice: Decimal
+    public let priceChangePercentage24H: Decimal
 }
